requirements.txt
pandas>=1.5.0
numpy>=1.21.0
faker>=19.0.0
streamlit>=1.28.0
flask>=2.3.0
openai>=1.0.0
anthropic>=0.25.0
python-dotenv>=1.0.0
plotly>=5.15.0
sqlalchemy>=2.0.0